Preparatory Makefile cleanups for packaging.
Use CPPFLAGS if specified (Debian uses CPPFLAGS to provide some hardening defines)
Use DESTDIR if specified (packaging systems use this to install to a temporary location when building)

Merge all the duplicated makefiles into one that builds all the binaries.
Stub out the old makefiles to delegate to the main one for the sake of people building with old instructions.