Use the standard init.d skeleton. Add support for running as a non-root user.
This leaves the standard init.d script untouched for now (it's not immediately
obvious how to do the run-as-user part there given that we can't create the
user). The Debian copy is now in debian/fadump1090.init and is based on the
Debian skeleton.

/etc/default/fadump1090 is read for options (command line args to dump1090,
whether to start on boot) and the user to run as.

By default, postinst creates a system user "fa1090" and the default defaults
match this.

Rely on debhelper to get the init scripts in the right place and to
handle start/stop of the service in postinst/prerm.